+++ core/classes/DTO/UserProfileField.php
@@ -4,7 +4,7 @@
  *
  * @package NamelessMC\DTO
  * @author Aberdeener
- * @version 2.0.0-pr13
+ * @version 2.2.0
  * @license MIT
  */
 class UserProfileField extends ProfileField
@@ -19,6 +19,12 @@ class UserProfileField extends ProfileField
         $this->value = $row->value;
         $this->updated = $row->updated;
         $this->upf_id = $row->upf_id;
+    }
+
+    public function purifyValue(): ?string
+    {
+        // TODO: option for field to support HTML
+        return Output::getClean($this->value);
     }
 
     public function updated()