+++ core/classes/Misc/Text.php
@@ -177,6 +177,9 @@ class Text
 
         $content = html_entity_decode($content);
 
+        // Also strip any mentions
+        $content = MentionsHook::stripPost(['content' => $content])['content'];
+
         return self::truncate($content, 512, [
             'html' => true,
         ]);