+++ core/includes/updates/212.php
@@ -5,9 +5,6 @@ return new class() extends UpgradeScript {
     {
         $this->runMigrations();
 
-        PurgeExpiredSessions::schedule(new Language('core', 'en_UK'));
-        PurgeInactiveUsers::schedule(new Language('core', 'en_UK'));
-
-        $this->setVersion('2.2.0');
+        $this->setVersion('2.1.3');
     }
 };