+++ modules/Core/pages/forgot_password.php
@@ -29,7 +29,7 @@ if ($user->isLoggedIn()) {
     Redirect::to(URL::build('/'));
 }
 
-if (!isset($_GET['c'])) {
+if (empty($_GET['c'])) {
     // Enter email address form
     if (Input::exists()) {
         if (Token::check()) {